Move window_all_wheel_input to libopenrct2ui
