+++ src/openrct2/ui/WindowManager.h
@@ -43,6 +43,7 @@ namespace OpenRCT2::Ui
         virtual void HandleKeyboard(bool isTitle) abstract;
         virtual std::string GetKeyboardShortcutString(sint32 shortcut) abstract;
         virtual void SetMainView(sint32 x, sint32 y, sint32 zoom, sint32 rotation) abstract;
+        virtual void UpdateMouseWheel() abstract;
     };
 
     IWindowManager * CreateDummyWindowManager();
